'''Refactor DistributedExecutionContext and call all callbacks from the main thread'''
 * DistributedExecutionContext: inherit from ManagerNetworkClientCallback
 * DistributedExecutionContext: Rename hostName and port into remoteHostName and remotePort; add 's' to numSentWorkUnit
 * Intelua: check intelua directory and make readable error if the directory is not correct
 * Refactor WorkUnitPool by splitting it into 4 classes:
   * WorkUnitPool: abstract base class
   * SingleWorkUnitPool: pool for a single work unit
   * CompositeWorkUnitPool: pool for a composite work unit
   * AsynchroneousWorkUnitPool: pool for asynchroneous work units
 * AsynchroneousWorkUnitPool: queue for finished work units, call callback from the calling thread
 * New lua function: ExecutionContext::waitUntilAllWorkUnitsAreDone()
 * Add timeOutInMillisecond to ExecutionContext::waitUntilAllWorkUnitsAreDone()
 * New virtual function: ExecutionContext::flushCallbacks()

+++ src/UserInterface/Component/ContainerCurveEditor.cpp
@@ -529,7 +529,7 @@ public:
     yAxisLabel->setBoundsRelative(0.35f, 0, 0.3f, 0.5f);
     yAxisLabel->setSize(yAxisLabel->getWidth(), 20);
     yListBox->setBoundsRelative(0.35f, 0.22f, 0.3f, 1.f);
-    yListBox->setSize(yListBox->getWidth(), yListBox->getHeight() - 0.22f * getHeight());
+    yListBox->setSize(yListBox->getWidth(), (int)(yListBox->getHeight() - 0.22f * getHeight()));
 
     //selectedCurves->setBoundsRelative(0.4f, 0, 0.6f, 1.f);
     selectAllButton->setBoundsRelative(0.7f, 0.22f, 0.3f, 0.5f);
